requirements.txt
django==4.2.9
environs[django]==11
pillow==11
django-debug-toolbar==4.4.6
django-phonenumber-field[phonenumbers]==8.0.0
djangorestframework==3.15.2
